7.1 Make utils for google gson json conversions and few StringUtils

+++ app/src/main/java/com/example/android/ecommerce/model/Category.kt
@@ -16,7 +16,7 @@ import kotlinx.android.parcel.Parcelize
 data class Category(
     @TypeConverters(ChildCategoryListTypeConverter::class)
     @SerializedName("child_categories")
-    var childCategories: MutableList<Long?>? = null,
+    var childCategories: MutableList<Int?>? = null,
     @SerializedName("id")
     @PrimaryKey (autoGenerate = true)
     var id: Long = 0, // 13
